Fixed a bunch of issues from code review:
* Store settings as CDATA in schema
* Make `key-binder` agnostic to the current applet shortcut
* Decouple Window settings from Applet settings
* Fix memory leak
* Fix settings callback to prevent segfault
